feat(design): rename DaffSizeable interface to DaffSizable
